refactor the batch upload, moved functions to utils package, created more tests
