+++ internal/loader.go
@@ -1,22 +1,15 @@
 package internal
 
 import (
-	"bufio"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
-	"runtime"
-	"strconv"
 	"strings"
-	"sync"
-	"time"
 
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
 	jsoniter "github.com/json-iterator/go"
-	"github.com/rs/zerolog/log"
 
 	"github.com/rtlnl/data-personalization-api/models"
 	"github.com/rtlnl/data-personalization-api/pkg/db"
@@ -53,23 +46,6 @@ type StreamingResponse struct {
 	Message string `json:"message"`
 }
 
-// Check if it is required to check the signal format
-func requireSignalFormat(m *models.Model) bool {
-	if len(m.SignalOrder) > 1 && m.Concatenator != "" {
-		return true
-	}
-	return false
-}
-
-// Check that the signal format is correct
-func correctSignalFormat(m *models.Model, s string) bool {
-	res := strings.FieldsFunc(s, func(c rune) bool {
-		r := []rune(m.Concatenator)
-		return c == r[0]
-	})
-	return len(m.SignalOrder) == len(res)
-}
-
 // CreateStreaming creates a new record in the selected campaign
 func CreateStreaming(c *gin.Context) {
 	ac := c.MustGet("AerospikeClient").(*db.AerospikeClient)
@@ -92,7 +68,7 @@ func CreateStreaming(c *gin.Context) {
 		return
 	}
 
-	if requireSignalFormat(m) && !correctSignalFormat(m, sr.Signal) {
+	if m.RequireSignalFormat() && !m.CorrectSignalFormat(sr.Signal) {
 		utils.ResponseError(c, http.StatusBadRequest, errors.New("the expected signal format must be "+strings.Join(m.SignalOrder, m.Concatenator)))
 		return
 	}
@@ -237,13 +213,14 @@ func Batch(c *gin.Context) {
 	}
 
 	// upload data from request itself
+	bo := NewBatchOperator(ac, m)
 	if len(br.Data) > 0 && br.Data != nil {
-		de, err := uploadDataDirectly(ac, br.Data, m)
+		ln, due, err := bo.UploadDataDirectly(br.Data)
 		if err != nil {
 			utils.ResponseError(c, http.StatusInternalServerError, err)
 			return
 		}
-		utils.Response(c, http.StatusCreated, de)
+		utils.Response(c, http.StatusCreated, &BatchResponse{NumberOfLines: ln, ErrorRecords: due})
 		return
 	}
 
@@ -254,7 +231,7 @@ func Batch(c *gin.Context) {
 	}
 
 	// upload data from S3 file
-	bucket, key := StripS3URL(br.DataLocation)
+	bucket, key := utils.StripS3URL(br.DataLocation)
 	s := db.NewS3Client(&db.S3Bucket{Bucket: bucket, ACL: ""}, sess)
 
 	// check if file exists
@@ -273,211 +250,9 @@ func Batch(c *gin.Context) {
 	batchID := uuid.New().String()
 
 	// seprate thread
-	go uploadDataFromFile(ac, f, m, batchID)
+	go bo.UploadDataFromFile(f, batchID)
 
 	utils.Response(c, http.StatusCreated, &BatchBulkResponse{BatchID: batchID})
-}
-
-// StripS3URL returns the bucket and the key from a s3 url location
-func StripS3URL(URL string) (string, string) {
-	bucketTmp := strings.Replace(URL, "s3://", "", -1)
-
-	bucket := bucketTmp[:strings.IndexByte(bucketTmp, '/')]
-	key := strings.TrimPrefix(URL, fmt.Sprintf("s3://%s/", bucket))
-
-	return bucket, key
-}
-
-func uploadDataDirectly(ac *db.AerospikeClient, bd []BatchData, m *models.Model) (*BatchResponse, error) {
-	var ln, ne int = 0, 0
-	var vl bool = false
-	var lineErrors []models.LineError
-
-	setName := m.ComposeSetName()
-
-	// check upfront if signal validation is required
-	if requireSignalFormat(m) {
-		vl = true
-	}
-
-	for _, data := range bd {
-		for sig, recommendedItems := range data {
-			ln++
-
-			// validate if required
-			if vl && !correctSignalFormat(m, sig) {
-				ne++
-				if ln <= maxErrorLines {
-					msg := fmt.Sprintf("wrong format, the expected signal format must be %s", strings.Join(m.SignalOrder, m.Concatenator))
-					lineErrors = append(lineErrors, models.LineError{strconv.Itoa(ln): msg})
-				}
-				continue
-			}
-
-			// upload to Aerospike
-			if err := ac.AddOne(setName, sig, binKey, recommendedItems); err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	return &BatchResponse{
-		NumberOfLines: strconv.Itoa(ln),
-		ErrorRecords: DataUploadedError{
-			Errors:              lineErrors,
-			NumberOfLinesFailed: strconv.Itoa(ne),
-		},
-	}, nil
-}
-
-// BulkStatus defines the status of the batch upload from S3
-type BulkStatus string
-
-const (
-	// BulkUploading represents the uploading status
-	BulkUploading = "UPLOADING"
-	// BulkSucceeded represents the succeeded status
-	BulkSucceeded = "SUCCEEDED"
-	// BulkPartialUpload represent the partial upload status
-	BulkPartialUpload = "PARTIAL UPLOAD"
-	// BulkFailed represents the failed status
-	BulkFailed = "FAILED"
-)
-
-var (
-	// MaxNumberOfWorkers is the max number of concurrent goroutines for uploading data
-	MaxNumberOfWorkers = runtime.NumCPU()
-)
-
-func uploadDataFromFile(ac *db.AerospikeClient, file *io.ReadCloser, m *models.Model, batchID string) {
-	start := time.Now()
-
-	// write to Aerospike it's uploading
-	if err := ac.AddOne(bulkStatusSetName, batchID, statusBinKey, BulkUploading); err != nil {
-		log.Panic().Msg(err.Error())
-	}
-
-	setName := m.ComposeSetName()
-	rd := bufio.NewReader(*file)
-	rs := make(chan *models.RecordQueue)
-	le := make(chan models.LineError)
-
-	// create sync group
-	wg := &sync.WaitGroup{}
-
-	// fillup the channel with lines
-	go func() {
-		iterateFile(rd, setName, m, rs, le)
-		close(rs)
-		close(le)
-	}()
-
-	// store eventual errors
-	go func() {
-		if storeErrors(ac, batchID, le) > 0 {
-			// write to Aerospike it partially uploded the data
-			if err := ac.AddOne(bulkStatusSetName, batchID, statusBinKey, BulkPartialUpload); err != nil {
-				log.Panic().Msg(err.Error())
-			}
-		} else {
-			// write to Aerospike it succeeded
-			if err := ac.AddOne(bulkStatusSetName, batchID, statusBinKey, BulkSucceeded); err != nil {
-				log.Panic().Msg(err.Error())
-			}
-		}
-	}()
-
-	// consumes all the lines in parallel based on number of cpus
-	wg.Add(MaxNumberOfWorkers)
-	for i := 0; i < MaxNumberOfWorkers; i++ {
-		go func() {
-			uploadRecord(ac, batchID, rs)
-			wg.Done()
-		}()
-	}
-
-	// wait until done
-	wg.Wait()
-
-	elapsed := time.Since(start)
-	log.Info().Msgf("Uploading took %s", elapsed)
-}
-
-func iterateFile(rd *bufio.Reader, setName string, m *models.Model, rs chan<- *models.RecordQueue, le chan<- models.LineError) {
-	var ln int = 0
-	var vl bool = false
-
-	// check upfront if signal validation is required
-	if requireSignalFormat(m) {
-		vl = true
-	}
-
-	eof := false
-	for !eof {
-		line, err := rd.ReadString('\n')
-		if err == io.EOF {
-			eof = true
-		}
-
-		// string new-line character
-		l := strings.TrimSuffix(line, "\n")
-
-		// marshal the object
-		var entry models.SingleEntry
-		if err := json.Unmarshal([]byte(l), &entry); err != nil {
-			le <- models.LineError{
-				"lineRaw": line,
-				"message": err.Error(),
-			}
-			continue
-		}
-
-		// validate signal format
-		ln++
-		if vl && !correctSignalFormat(m, entry.SignalID) {
-			le <- models.LineError{
-				"line":    strconv.Itoa(ln),
-				"message": "signal not formatted correctly",
-			}
-			continue
-		}
-
-		// add to channel
-		rs <- &models.RecordQueue{SetName: setName, Entry: entry, Error: nil}
-	}
-}
-
-func uploadRecord(ac *db.AerospikeClient, batchID string, rs chan *models.RecordQueue) {
-	// upload record to aerospike when it arrives
-	for r := range rs {
-		if err := ac.AddOne(r.SetName, r.Entry.SignalID, binKey, r.Entry.Recommended); err != nil {
-			// write to Aerospike it failed
-			if err := ac.AddOne(bulkStatusSetName, batchID, statusBinKey, BulkFailed); err != nil {
-				// if this fails than since we cannot return the request to the user
-				// we need to restart the application
-				log.Panic().Msg(err.Error())
-			}
-			return
-		}
-	}
-}
-
-func storeErrors(ac *db.AerospikeClient, batchID string, le chan models.LineError) int {
-	allErrors := []models.LineError{}
-	i := 0
-	for lineError := range le {
-		if i < maxErrorLines {
-			allErrors = append(allErrors, lineError)
-			i++
-			continue
-		}
-		break
-	}
-	// save to Aerospike the errors list
-	if err := ac.AddOne(bulkStatusSetName, batchID, lineBinError, allErrors); err != nil {
-		log.Panic().Msg(err.Error())
-	}
-	return i
 }
 
 // BatchStatusResponse is the response payload for getting the status of the bulk upload from S3
@@ -511,8 +286,8 @@ func BatchStatus(c *gin.Context) {
 	}
 }
 
-// The objects coming from Aerospike have type []interface{}. This function converts
-// the Bins in the appropriate type for consistency
+// convertLineError converts the objects coming from Aerospike have type []interface{}.
+// This function converts the Bins in the appropriate type for consistency
 func convertLineError(bins interface{}) []models.LineError {
 	var linesError []models.LineError
 	newBins := bins.([]interface{})
